LayoutTests/inspector/model/source-map-spec-expected.txt

Run source map specification consumer test cases.


== Running test suite: SourceMapSpec
-- Running test case: SourceMapSpec

TEST CASE: versionValid (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.

TEST CASE: versionMissing (expect INVALID)
WARN: Source Map "version-missing.js.map" has invalid "version"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: versionNotANumber (expect INVALID)
WARN: Source Map "version-not-a-number.js.map" has invalid "version"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: versionNumericString (expect INVALID)
WARN: Source Map "version-numeric-string.js.map" has invalid "version"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: versionTooHigh (expect INVALID)
WARN: Source Map "version-too-high.js.map" has invalid "version"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: versionTooLow (expect INVALID)
WARN: Source Map "version-too-low.js.map" has invalid "version"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: mappingsMissing (expect INVALID)
WARN: Source Map "mappings-missing.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: sourcesMissing (expect INVALID)
WARN: Source Map "sources-missing.js.map" has invalid "sources"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: sourcesNotAList1 (expect INVALID)
WARN: Source Map "sources-not-a-list-1.js.map" has invalid "sources"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: sourcesNotAList2 (expect INVALID)
WARN: Source Map "sources-not-a-list-2.js.map" has invalid "sources"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: sourcesNotStringOrNull (expect INVALID)
WARN: Source Map "sources-not-string-or-null.js.map" has invalid "sources"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: sourcesContentMissing (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.

TEST CASE: sourcesContentNotAList1 (expect INVALID)
WARN: Source Map "sources-content-not-a-list-1.js.map" has invalid "sourcesContent"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: sourcesContentNotAList2 (expect INVALID)
WARN: Source Map "sources-content-not-a-list-2.js.map" has invalid "sourcesContent"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: sourcesContentNotStringOrNull (expect INVALID)
WARN: Source Map "sources-not-string-or-null.js.map" has invalid "sources"
FAIL: Script resource should have loaded
    Expected: truthy
    Actual: undefined

TEST CASE: sourcesAndSourcesContentBothNull (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.

TEST CASE: fileNotAString1 (expect INVALID)
WARN: Source Map "file-not-a-string-1.js.map" has invalid "file"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: fileNotAString2 (expect INVALID)
WARN: Source Map "file-not-a-string-2.js.map" has invalid "file"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: sourceRootNotAString1 (expect INVALID)
WARN: Source Map "source-root-not-a-string-1.js.map" has invalid "sourceRoot"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: sourceRootNotAString2 (expect INVALID)
WARN: Source Map "source-root-not-a-string-2.js.map" has invalid "sourceRoot"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: namesMissing (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.

TEST CASE: namesNotAList1 (expect INVALID)
WARN: Source Map "names-not-a-list-1.js.map" has invalid "names"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: namesNotAList2 (expect INVALID)
WARN: Source Map "names-not-a-list-2.js.map" has invalid "names"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: namesNotString (expect INVALID)
WARN: Source Map "names-not-string.js.map" has invalid "names"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: ignoreListEmpty (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.

TEST CASE: ignoreListValid1 (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Ignored sources should be in ignoreList

TEST CASE: ignoreListWrongType1 (expect INVALID)
WARN: Source Map "ignore-list-wrong-type-1.js.map" has invalid "ignoreList"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: ignoreListWrongType2 (expect INVALID)
WARN: Source Map "ignore-list-wrong-type-2.js.map" has invalid "ignoreList"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: ignoreListWrongType3 (expect INVALID)
WARN: Source Map "ignore-list-wrong-type-3.js.map" has invalid "ignoreList"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: ignoreListWrongType4 (expect INVALID)
WARN: Source Map "ignore-list-wrong-type-4.js.map" has invalid "ignoreList"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: ignoreListOutOfBounds1 (expect INVALID)
WARN: Source Map "ignore-list-out-of-bounds-1.js.map" has invalid "ignoreList"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: ignoreListOutOfBounds2 (expect INVALID)
WARN: Source Map "ignore-list-out-of-bounds-2.js.map" has invalid "ignoreList"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: unrecognizedProperty (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.

TEST CASE: invalidVLQDueToNonBase64Character (expect INVALID)
WARN: Source Map "invalid-vlq-non-base64-char.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidVLQDueToNonBase64CharacterPadding (expect INVALID)
WARN: Source Map "invalid-vlq-non-base64-char-padding.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidVLQDueToMissingContinuationDigits (expect INVALID)
WARN: Source Map "invalid-vlq-missing-continuation.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingNotAString1 (expect INVALID)
WARN: Source Map "invalid-mapping-not-a-string-1.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingNotAString2 (expect INVALID)
WARN: Source Map "invalid-mapping-not-a-string-2.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentBadSeparator (expect INVALID)
WARN: Source Map "invalid-mapping-bad-separator.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithZeroFields (expect INVALID)
WARN: Source Map "invalid-mapping-segment-with-zero-fields.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithTwoFields (expect INVALID)
WARN: Source Map "invalid-mapping-segment-with-two-fields.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithThreeFields (expect INVALID)
WARN: Source Map "invalid-mapping-segment-with-three-fields.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithSourceIndexOutOfBounds (expect INVALID)
WARN: Source Map "invalid-mapping-segment-source-index-out-of-bounds.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithNameIndexOutOfBounds (expect INVALID)
WARN: Source Map "invalid-mapping-segment-name-index-out-of-bounds.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithNegativeColumn (expect INVALID)
WARN: Source Map "invalid-mapping-segment-negative-column.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithNegativeSourceIndex (expect INVALID)
WARN: Source Map "invalid-mapping-segment-negative-source-index.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithNegativeOriginalLine (expect INVALID)
WARN: Source Map "invalid-mapping-segment-negative-original-line.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithNegativeOriginalColumn (expect INVALID)
WARN: Source Map "invalid-mapping-segment-negative-original-column.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithNegativeNameIndex (expect INVALID)
WARN: Source Map "invalid-mapping-segment-negative-name-index.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithNegativeRelativeColumn (expect INVALID)
WARN: Source Map "invalid-mapping-segment-negative-relative-column.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithNegativeRelativeSourceIndex (expect INVALID)
WARN: Source Map "invalid-mapping-segment-negative-relative-source-index.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithNegativeRelativeOriginalLine (expect INVALID)
WARN: Source Map "invalid-mapping-segment-negative-relative-original-line.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithNegativeRelativeOriginalColumn (expect INVALID)
WARN: Source Map "invalid-mapping-segment-negative-relative-original-column.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithNegativeRelativeNameIndex (expect INVALID)
WARN: Source Map "invalid-mapping-segment-negative-relative-name-index.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithColumnExceeding32Bits (expect INVALID)
WARN: Source Map "invalid-mapping-segment-column-too-large.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithSourceIndexExceeding32Bits (expect INVALID)
WARN: Source Map "invalid-mapping-segment-source-index-too-large.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithOriginalLineExceeding32Bits (expect INVALID)
WARN: Source Map "invalid-mapping-segment-original-line-too-large.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithOriginalColumnExceeding32Bits (expect INVALID)
WARN: Source Map "invalid-mapping-segment-original-column-too-large.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: invalidMappingSegmentWithNameIndexExceeding32Bits (expect INVALID)
WARN: Source Map "invalid-mapping-segment-name-index-too-large.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: validMappingFieldsWith32BitMaxValues (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.

TEST CASE: validMappingLargeVLQ (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.

TEST CASE: validMappingEmptyGroups (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.

TEST CASE: validMappingEmptyString (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.

TEST CASE: indexMapWrongTypeSections (expect INVALID)
WARN: Source Map "index-map-wrong-type-sections.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapWrongTypeOffset (expect INVALID)
WARN: Source Map "index-map-wrong-type-offset.js.map" has invalid "offset"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapWrongTypeMap (expect INVALID)
WARN: Source Map "index-map-wrong-type-map.js.map" has invalid "map"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapInvalidBaseMappings (expect INVALID)
WARN: Source Map "index-map-invalid-base-mappings.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapInvalidOverlap (expect INVALID)
WARN: Source Map "index-map-invalid-overlap.js.map" has invalid "offset.column"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapInvalidOrder (expect INVALID)
WARN: Source Map "index-map-invalid-order.js.map" has invalid "offset.line"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapMissingMap (expect INVALID)
WARN: Source Map "index-map-missing-map.js.map" has invalid "map"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapInvalidSubMap (expect INVALID)
WARN: Source Map "index-map-invalid-sub-map.js.map" has invalid "mappings"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapMissingOffset (expect INVALID)
WARN: Source Map "index-map-missing-offset.js.map" has invalid "offset"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapMissingOffsetLine (expect INVALID)
WARN: Source Map "index-map-missing-offset-line.js.map" has invalid "offset.line"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapMissingOffsetColumn (expect INVALID)
WARN: Source Map "index-map-missing-offset-column.js.map" has invalid "offset.column"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapOffsetLineWrongType (expect INVALID)
WARN: Source Map "index-map-offset-line-wrong-type.js.map" has invalid "offset.line"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapOffsetColumnWrongType (expect INVALID)
WARN: Source Map "index-map-offset-column-wrong-type.js.map" has invalid "offset.column"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapEmptySections (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.

TEST CASE: indexMapFileWrongType1 (expect INVALID)
WARN: Source Map "index-map-file-wrong-type-1.js.map" has invalid "file"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: indexMapFileWrongType2 (expect INVALID)
WARN: Source Map "index-map-file-wrong-type-2.js.map" has invalid "file"
PASS: Script resource should have loaded
PASS: Expected that there is an associated failed source map URL
PASS: Expected no source map resource loaded

TEST CASE: basicMapping (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 0) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 9) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 15) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 2, expected: 2
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 22) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 24) should be mapped
PASS: Original line: 2, expected: 2
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 25) should be mapped
PASS: Original line: 3, expected: 3
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 34) should be mapped
PASS: Original line: 3, expected: 3
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 40) should be mapped
PASS: Original line: 4, expected: 4
PASS: Original column: 2, expected: 2
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 47) should be mapped
PASS: Original line: 4, expected: 4
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 49) should be mapped
PASS: Original line: 5, expected: 5
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 50) should be mapped
PASS: Original line: 6, expected: 6
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 56) should be mapped
PASS: Original line: 7, expected: 7
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js

TEST CASE: sourceRootResolution (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 0) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 0, expected: 0
FAIL: Original source: basic-mapping-original.js, expected: theroot/basic-mapping-original.js
    Expected: "theroot/basic-mapping-original.js"
    Actual: "basic-mapping-original.js"
PASS: Test location (0, 9) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 9, expected: 9
FAIL: Original source: basic-mapping-original.js, expected: theroot/basic-mapping-original.js
    Expected: "theroot/basic-mapping-original.js"
    Actual: "basic-mapping-original.js"

TEST CASE: sourceResolutionAbsoluteURL (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 0) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 0, expected: 0
FAIL: Original source: basic-mapping-original.js, expected: /baz/quux/basic-mapping-original.js
    Expected: "/baz/quux/basic-mapping-original.js"
    Actual: "basic-mapping-original.js"
PASS: Test location (0, 9) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 9, expected: 9
FAIL: Original source: basic-mapping-original.js, expected: /baz/quux/basic-mapping-original.js
    Expected: "/baz/quux/basic-mapping-original.js"
    Actual: "basic-mapping-original.js"

TEST CASE: basicMappingWithIndexMap (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 0) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 9) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 15) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 2, expected: 2
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 22) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 24) should be mapped
PASS: Original line: 2, expected: 2
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 25) should be mapped
PASS: Original line: 3, expected: 3
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 34) should be mapped
PASS: Original line: 3, expected: 3
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 40) should be mapped
PASS: Original line: 4, expected: 4
PASS: Original column: 2, expected: 2
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 47) should be mapped
PASS: Original line: 4, expected: 4
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 49) should be mapped
PASS: Original line: 5, expected: 5
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 50) should be mapped
PASS: Original line: 6, expected: 6
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 56) should be mapped
PASS: Original line: 7, expected: 7
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js

TEST CASE: indexMapWithMissingFile (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 0) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 9) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 15) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 2, expected: 2
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 22) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 24) should be mapped
PASS: Original line: 2, expected: 2
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 25) should be mapped
PASS: Original line: 3, expected: 3
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 34) should be mapped
PASS: Original line: 3, expected: 3
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 40) should be mapped
PASS: Original line: 4, expected: 4
PASS: Original column: 2, expected: 2
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 47) should be mapped
PASS: Original line: 4, expected: 4
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 49) should be mapped
PASS: Original line: 5, expected: 5
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 50) should be mapped
PASS: Original line: 6, expected: 6
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 56) should be mapped
PASS: Original line: 7, expected: 7
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js

TEST CASE: indexMapWithTwoConcatenatedSources (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 0) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 9) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 15) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 2, expected: 2
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 22) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 24) should be mapped
PASS: Original line: 2, expected: 2
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 25) should be mapped
PASS: Original line: 3, expected: 3
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 34) should be mapped
PASS: Original line: 3, expected: 3
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 40) should be mapped
PASS: Original line: 4, expected: 4
PASS: Original column: 2, expected: 2
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 47) should be mapped
PASS: Original line: 4, expected: 4
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 49) should be mapped
PASS: Original line: 5, expected: 5
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 50) should be mapped
PASS: Original line: 6, expected: 6
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 56) should be mapped
PASS: Original line: 7, expected: 7
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 62) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 0, expected: 0
PASS: Original source: second-source-original.js, expected: second-source-original.js
PASS: Test location (0, 71) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 9, expected: 9
PASS: Original source: second-source-original.js, expected: second-source-original.js
PASS: Test location (0, 77) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 2, expected: 2
PASS: Original source: second-source-original.js, expected: second-source-original.js
PASS: Test location (0, 83) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 9, expected: 9
PASS: Original source: second-source-original.js, expected: second-source-original.js
PASS: Test location (0, 88) should be mapped
PASS: Original line: 2, expected: 2
PASS: Original column: 0, expected: 0
PASS: Original source: second-source-original.js, expected: second-source-original.js
PASS: Test location (0, 89) should be mapped
PASS: Original line: 3, expected: 3
PASS: Original column: 0, expected: 0
PASS: Original source: second-source-original.js, expected: second-source-original.js

TEST CASE: sourcesNullSourcesContentNonNull (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 0) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 0, expected: 0
PASS: Original source: null, expected: null
PASS: Test location (0, 9) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 9, expected: 9
PASS: Original source: null, expected: null

TEST CASE: sourcesNonNullSourcesContentNull (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 0) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 0, expected: 0
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js
PASS: Test location (0, 9) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 9, expected: 9
PASS: Original source: basic-mapping-original.js, expected: basic-mapping-original.js

TEST CASE: transitiveMapping (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
Transitive mapping test ignored
Transitive mapping test ignored
Transitive mapping test ignored
Transitive mapping test ignored
Transitive mapping test ignored
Transitive mapping test ignored
Transitive mapping test ignored
Transitive mapping test ignored

TEST CASE: transitiveMappingWithThreeSteps (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
Transitive mapping test ignored
Transitive mapping test ignored
Transitive mapping test ignored
Transitive mapping test ignored
Transitive mapping test ignored
Transitive mapping test ignored
Transitive mapping test ignored
Transitive mapping test ignored

TEST CASE: vlqValidSingleDigit (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 15) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 0, expected: 0
PASS: Original source: vlq-valid-single-digit-original.js, expected: vlq-valid-single-digit-original.js

TEST CASE: vlqValidNegativeDigit (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (2, 15) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 3, expected: 3
PASS: Original source: vlq-valid-negative-digit-original.js, expected: vlq-valid-negative-digit-original.js
PASS: Test location (2, 2) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 1, expected: 1
PASS: Original source: vlq-valid-negative-digit-original.js, expected: vlq-valid-negative-digit-original.js

TEST CASE: vlqValidContinuationBitPresent1 (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 15) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 1, expected: 1
PASS: Original source: vlq-valid-continuation-bit-present-1-original.js, expected: vlq-valid-continuation-bit-present-1-original.js

TEST CASE: vlqValidContinuationBitPresent2 (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (2, 16) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 1, expected: 1
PASS: Original source: vlq-valid-continuation-bit-present-2-original.js, expected: vlq-valid-continuation-bit-present-2-original.js

TEST CASE: mappingSemanticsSingleFieldSegment (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 0) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 1, expected: 1
PASS: Original source: mapping-semantics-single-field-segment-original.js, expected: mapping-semantics-single-field-segment-original.js
FAIL: Test location (0, 2) should be mapped
    Expected: truthy
    Actual: false
FAIL: Original line: 0, expected: null
    Expected: null
    Actual: 0
FAIL: Original column: 2, expected: null
    Expected: null
    Actual: 2
FAIL: Original source: mapping-semantics-single-field-segment.js, expected: null
    Expected: null
    Actual: "mapping-semantics-single-field-segment.js"

TEST CASE: mappingSemanticsFourFieldSegment (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 1) should be mapped
PASS: Original line: 2, expected: 2
PASS: Original column: 2, expected: 2
PASS: Original source: mapping-semantics-four-field-segment-original.js, expected: mapping-semantics-four-field-segment-original.js

TEST CASE: mappingSemanticsFiveFieldSegment (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 1) should be mapped
PASS: Original line: 2, expected: 2
PASS: Original column: 2, expected: 2
PASS: Original source: mapping-semantics-five-field-segment-original.js, expected: mapping-semantics-five-field-segment-original.js

TEST CASE: mappingSemanticsColumnReset (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 1) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 0, expected: 0
PASS: Original source: mapping-semantics-column-reset-original.js, expected: mapping-semantics-column-reset-original.js
PASS: Test location (1, 1) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 0, expected: 0
PASS: Original source: mapping-semantics-column-reset-original.js, expected: mapping-semantics-column-reset-original.js

TEST CASE: mappingSemanticsRelative1 (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 1) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 0, expected: 0
PASS: Original source: mapping-semantics-relative-1-original.js, expected: mapping-semantics-relative-1-original.js
PASS: Test location (0, 5) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 4, expected: 4
PASS: Original source: mapping-semantics-relative-1-original.js, expected: mapping-semantics-relative-1-original.js

TEST CASE: mappingSemanticsRelative2 (expect VALID)
PASS: Script resource and source map event should have triggered
PASS: Resource should have loaded 1 SourceMap.
PASS: SourceMap should be a WI.SourceMap instance.
PASS: Test location (0, 1) should be mapped
PASS: Original line: 0, expected: 0
PASS: Original column: 2, expected: 2
PASS: Original source: mapping-semantics-relative-2-original.js, expected: mapping-semantics-relative-2-original.js
PASS: Test location (1, 2) should be mapped
PASS: Original line: 1, expected: 1
PASS: Original column: 2, expected: 2
PASS: Original source: mapping-semantics-relative-2-original.js, expected: mapping-semantics-relative-2-original.js
